# swagger.yaml
openapi: 3.1.0
info:
  title: T-shirt Store Intenship Project
  description: Creating an API for e-commerce
  version: 1.1.0
  contact: 
    email: ariffidan03@gmail.com
    url: "https://scramblebit.com/"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - regular
servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local host - secure
    variables: 
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http
components: 
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /userdashboard/update:
    put:
      summary: Update User Information
      description: Allows a user to update their name, email, and profile picture.
      operationId: updateUser
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the user.
                email:
                  type: string
                  description: The email address of the user.
                photo:
                  type: string
                  format: binary
                  description: The new profile photo to upload (optional).
      responses:
        '200':
          description: User successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        description: The user's unique ID
                      name:
                        type: string
                        description: The updated name of the user
                      email:
                        type: string
                        description: The updated email of the user
                      photo:
                        type: object
                        properties:
                          id:
                            type: string
                            description: The Cloudinary photo ID
                          secure_url:
                            type: string
                            description: The secure URL of the uploaded photo
              '400':
                description: Bad request - Missing required fields or invalid file
              '401':
                description: Unauthorized - Invalid or missing token
              '500':
                description: Internal server error
      components:
        securitySchemes:
          bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
  /admin/users/{id}:
    put:
      summary: Update a specific user's information
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User updated successfully

    delete:
      summary: Delete a specific user
      operationId: adminDeleteOneUser
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the user to delete
          schema:
            type: string
      responses:
        '200':
          description: User successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Invalid or missing token
        '404':
          description: User not found
        '500':
          description: Internal server error

# product part
...